feat: add ort benchmark script in xp folder
